% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_analysis.R
\name{covariate_ci}
\alias{covariate_ci}
\title{Compute transition probability matrices for plotting}
\usage{
covariate_ci(
  hmm,
  len_covariates,
  design,
  n = 100,
  level = 0.975,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{len_covariates}{A value indicating how many covariate values will be
used.}

\item{design}{A list of design matrices, one for each subject which
indicate the values of the each of the covariates (column) at each point
in time (row).}

\item{num_states}{The number of states in the desired HMM.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{conf_intervals}{A list of the confidence intervals for each fitted
parameter of the HMM as outputted by the functions \code{norm_ci}, \code{gam_ci}, or
\code{gam0_ci}.}
}
\value{
A list of 3 matrices, one for the estimate and the upper, and lower
confidence interval, with the rows of the matrices containing the entries
of the transition probability matrices at each covariate value.
}
\description{
This function computes the transition probability matrix entries with their
confidence intervals for different covariate values in order to plot how
they change. THIS FUNCTION DOES NOT SUPPORT UNPOOLED TRANSITION PROBABILITY
MATRICES.
}
