% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Compute Euclidean distance}
\usage{
euclidean_distance(x_positions, y_positions)
}
\arguments{
\item{x_positions}{A vector of positions along the x-axis.}

\item{y_positions}{A vector of positions along the y-axis.}
}
\value{
A vector of euclidean distances.
}
\description{
This function computes the euclidean distance between two vectors of
coordinates.
}
