% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tpm.R
\name{fit_tpm}
\alias{fit_tpm}
\title{Compute transition probability matrices}
\usage{
fit_tpm(num_states, num_subjects, num_time, beta, design)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_time}{A value indicating the length of the data.}

\item{beta}{A matrix of regression coefficients for the effect of the
covariates on the transition probability matrices \code{gamma}.}

\item{design}{A list of design matrices, one for each subject which
indicate the values of the each of the covariates (column) at each point
in time (row).}
}
\value{
A list containing a list for each subject/trial of the transition
probability matrices at each time step.
}
\description{
This function computes the time dependent transition probability matrix
from the corresponding vector \code{design_vec} and the matrix of regression
coefficients \code{beta}.
}
\examples{
# define parameters
num_states   <- 3
num_subjects <- 1
num_time     <- 1000
beta             <- matrix(c(0.01, 0.02, 0.001,
                             0.01, 0.03, 0.004,
                             0.01, 0.01, 0.003,
                             0.01, 0.04, 0.002,
                             0.01, 0.01, 0.004,
                             0.01, 0.03, 0.001),
                             ncol = 3, nrow = 6, byrow = TRUE)
design           <- list(matrix(0, nrow = 1000, ncol = 3))
design[[1]][, 1] <- 1 # First column is the intercept
design[[1]][, 2] <- sample(c(0, 1), size = 1000,
                           prob = c(0.3, 0.7), replace = TRUE)
design[[1]][, 3] <- rnorm(1000, mean = 5, sd = 1)

# compute the transition probability matrices
fit_tpm(num_states, num_subjects, num_time, beta, design)
}
