% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foralg}
\alias{foralg}
\title{Execute the forward algorithm}
\usage{
foralg(n, N, foo, gamma, allprobs)
}
\arguments{
\item{n}{The length of the timeseries.}

\item{N}{The number of states in the HMM.}

\item{foo}{The initial state distributions (delta).}

\item{gamma}{The transition probability matrix.}

\item{allprobs}{As outputted by the funtions norm_allporbs}
}
\description{
Execute the forward algorithm
}
