% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam0_allprobs.R
\name{gam0_allprobs}
\alias{gam0_allprobs}
\title{Compute allprobs}
\usage{
gam0_allprobs(
  num_states,
  num_variables,
  num_subjects,
  num_time,
  x,
  pn,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_time}{A value indicating the length of the data.}

\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{pn}{A list containing the natural parameters for the gamma HMM.}

\item{state_dep_dist_pooled}{A logical variable indicating whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}
}
\value{

}
\description{
This function computes the probability of each observation arising from each
state.
}
