% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam0_fit_hmm.R
\name{gam0_fit_hmm}
\alias{gam0_fit_hmm}
\title{Fit an HMM}
\usage{
gam0_fit_hmm(
  x,
  design,
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  alpha0,
  theta0,
  zweight0,
  beta0,
  delta0,
  state_dep_dist_pooled = FALSE,
  iterlim = 100,
  hessian = FALSE,
  print.level = 0
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{design}{A list of design matrices for each subject with each row
indicating the time and each column indicating the value of the
covariate.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{alpha0}{The starting values for the means of the gamma distributed
state dependent distributions of the HMM. \code{alpha0} is a list of matrices,
each matrix corresponding to a different variable in the data being fit.
The columns of the matrices correspond to the state number and the rows
correspond to the subject number.}

\item{theta0}{The starting values for the standard deviations of the
gamma distributed state dependent distributions of the HMM. \code{theta0}
is a list of matrices, each matrix corresponding to a different variable
in the data being fit. The columns of the matrices correspond to the
state number and the rows correspond to the subject number.}

\item{beta0}{A matrix of the initial regression coefficients for the effect
of the covariates on the transition probability matrices \code{gamma}.}

\item{delta0}{A list with each element being the starting initial state
distribution vector of the HMM for the subject corresponding to that
index.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}

\item{iterlim}{A value indicating the number of iterations \code{nlm()} should run
before exiting.}

\item{hessian}{A logical variable indicating whether to compute the hessian
at the minimum.}

\item{@param}{zweight0 The starting values for the weights of the zero
inflated portion of the gamma distributions. \code{zweight0} is a list of
vectors, one for each variable, and each vector element correspomnds to a
different subject.}
}
\value{
A list of parameters that specify the fitted gamma HMM.
}
\description{
This function fits data with an HMM by maximizing the likelihood estimate
given initial gamma parameters.
}
