% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_histogram_plotting.R
\name{gam0_hist_viterbi}
\alias{gam0_hist_viterbi}
\title{Plot histograms}
\usage{
gam0_hist_viterbi(
  x,
  viterbi,
  num_states,
  num_subjects,
  num_variables,
  hmm,
  state_dep_dist_pooled = FALSE,
  width = 1,
  n = 100,
  level = 0.975,
  x_step = 0.2
)
}
\value{
Histograms of the data with the distributions overlayed.
}
\description{
Plot histograms
}
