% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam0_ci_functions.R
\name{gam0_working_ind}
\alias{gam0_working_ind}
\title{Get indices of parameters in working vector}
\usage{
gam0_working_ind(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of the start and end indices.
}
\description{
This function finds the indices for each parameter of the zero inflated gamma
HMM within the vector of working parameters outputted by
\code{gam0_working_params()}.
}
