% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam0_working_params.R
\name{gam0_working_params}
\alias{gam0_working_params}
\title{Transform gamma HMM parameters from natural to working}
\usage{
gam0_working_params(
  num_states,
  num_variables,
  num_subjects,
  alpha,
  theta,
  zweight,
  beta,
  delta
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{alpha}{A list of matrices containing the shape parameters of the state
dependent gamma distributions. Each matrix corresponds to a different
variable, each row corresponds to a different subject and each column
corresponds to a different state.}

\item{theta}{A list of matrices containing the scale parameters of the
state dependent gamma distributions. Each matrix corresponds to a
different variable, each row corresponds to a different subject and each
column corresponds to a different state.}

\item{zweight}{A list of vectors, one for each variable, containing the
weights for the zero inflated gamma distribution.}

\item{beta}{A matrix of regression coefficients for the effect of the
covariates on the transition probability matrix \code{gamma}.}

\item{delta}{A list of the initial state distributions for each subject.}
}
\value{
A single vector containing working parameters.
}
\description{
The function transforms the natural gamma HMM parameters that have
additional constraints into working parameters that incorporate the
constraints. Zero inflated gamma funciton.
}
