% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_ci_functions.R
\name{gam_dist_ci_data}
\alias{gam_dist_ci_data}
\title{Compute the confidence intervals of fitted gamma distributions}
\usage{
gam_dist_ci_data(
  x,
  num_states,
  num_variables,
  num_subjects,
  sample,
  state_dep_dist_pooled = FALSE,
  x_step = 0.2,
  n = 100,
  level = 0.975
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{sample}{A list of \code{alpha} and \code{theta} sampled according to \code{gam_ci()}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}

\item{x_step}{A value indicating the step length for the range of
observation values.}

\item{n}{The number of samples in the Monte Carlo fitting.}

\item{level}{A number indicating the level of confidence for the desired
interval.}
}
\value{
A list containing the upper and lower confidence intervals and the
parameter esimates.
}
\description{
This is a helper function for \code{gam_hist_ci()} which computes the confidence
intervals for the fitted gamma state dependent distributions by utilizing
\code{gam_ci()}.
}
