% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_forecast_psr.R
\name{gam_forecast_psr}
\alias{gam_forecast_psr}
\title{Compute forecast gamma pseudo-residuals}
\usage{
gam_forecast_psr(x, hmm, state_dep_dist_pooled = FALSE)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{hmm}{A list of parameters that specify the gamma HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{alpha}, \code{theta}, \code{gamma},
\code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indicating whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}
}
\value{
A list of vectors (one for each subject) of the pseudo-residuals.
}
\description{
This function computes the gamma forecast pseudo-residuals of the data \code{x}
fitted with \code{hmm} (with gamma state dependent distributions).
}
