% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_ci_functions.R
\name{gam_hist_ci}
\alias{gam_hist_ci}
\title{Plot histograms with confidence intervals}
\usage{
gam_hist_ci(
  x,
  viterbi,
  num_states,
  num_subjects,
  num_variables,
  hmm,
  state_dep_dist_pooled = FALSE,
  width = 1,
  n = 100,
  level = 0.975,
  x_step = 0.2
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{viterbi}{A matrix with each column indicating the sequence of states
decoded by \code{gam_viterbi()} that is supposed to have generated the data of
the subject in the corresponding column.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_variables}{The number of variables in the data.}

\item{hmm}{A list of parameters that specify the gamma HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{alpha}, \code{theta}, \code{gamma},
\code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}

\item{width}{The width of the histogram bins.}

\item{n}{The number of samples in the Monte Carlo fitting.}

\item{level}{A number indicating the level of confidence for the desired
interval.}

\item{x_step}{A value indicating the step length for the range of
observation values.}
}
\value{
Histograms of the data with overlayed distributions and confidence
intervals.
}
\description{
This function plots the histograms for each subject and variable with the
fitted state dependent distributions overlayed and their corresponding
confidence intervals.
}
