% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_natural_params.R
\name{gam_natural_params}
\alias{gam_natural_params}
\title{Transform gamma HMM parameters from working to natural}
\usage{
gam_natural_params(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{working_params}{A vector of the working gamma parameters for the
HMM as outputted by \code{norm_working_params()}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}
}
\value{
A list of the natural paramters.
}
\description{
This function transforms the working gamma HMM parameters back into the
original format of the natural parameters and outputs them as a list. This
function is the reverse of \code{norm_working_params()}.
}
