% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_ci_functions.R
\name{gam_natural_vec}
\alias{gam_natural_vec}
\title{Convert working parameters to a vector of natural parameters}
\usage{
gam_natural_vec(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{working_params}{A vector of the working gamma parameters for the
HMM.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{alpha} and \code{theta} should be
treated as equal for all subjects.}
}
\value{
A vector of the natural parameters.
}
\description{
This funciton converts the vecgtor of working parameters to a vector of
natural parameters not inculding delta.
}
