% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm0_natural_params.R
\name{norm0_natural_params}
\alias{norm0_natural_params}
\title{Transform normal HMM parameters from working to natural}
\usage{
norm0_natural_params(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{working_params}{A vector of the working normal parameters for the
HMM as outputted by \code{norm_working_params()}.}

\item{state_dep_dist_pooled}{A logical variable indicating whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of the natural parameters.
}
\description{
This function transforms the working normal HMM parameters back into the
original format of the natural parameters and outputs them as a list. This
function is the reverse of \code{norm_working_params()}.
}
