% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_allprobs.R
\name{norm_allprobs}
\alias{norm_allprobs}
\title{Compute allprobs}
\usage{
norm_allprobs(
  num_states,
  num_variables,
  num_subjects,
  num_time,
  x,
  pn,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_time}{A value indicating the length of the data.}

\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{pn}{A list conatining the natural parameters for the normal HMM.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{

}
\description{
This function computes the probability of each observation arising from each
state.
}
