% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confindence_interval_functions.R
\name{norm_ci_data}
\alias{norm_ci_data}
\title{Reformat confidence interval data}
\usage{
norm_ci_data(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  estimate_vec,
  upper_vec,
  lower_vec,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability depends on.}

\item{estimate_vec}{A vector containing the estimated natural parameters.}

\item{upper_vec}{A vector containing the upper confidence interval of the
estimated natural parameters.}

\item{lower_vec}{A vector containing the lower confidence interval of the
estimated natural parameters.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list containing the upper and lower confidence interval for each
estimated parameter.
}
\description{
Reformat confidence interval data
}
