% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_forecast_psr.R
\name{norm_forecast_psr}
\alias{norm_forecast_psr}
\title{Compute forecast normal pseuso-residuals}
\usage{
norm_forecast_psr(x, hmm, state_dep_dist_pooled = FALSE)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{mu}, \code{sigma}, \code{gamma},
\code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of vectors (one for each subject) of the pseudo-residuals.
}
\description{
This function computes the normal forecast pseudo-residuals of the data \code{x}
fitted with \code{hmm} (with normal state dependent distributions).
}
