% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_gamma.R
\name{norm_gamma}
\alias{norm_gamma}
\title{Compute transition probability matrix}
\usage{
norm_gamma(num_states, beta, design_vec)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{beta}{A matrix of regression coefficients.}

\item{design_vec}{A vector from the design matrix with indicates the values
of the covariates.}
}
\value{
A matrix.
}
\description{
This function computes the time dependent transition probability matrix
from the corresponding vector \code{design_vec}  and the matrix of regression
coefficients \code{beta}.
}
\examples{
norm_gamma(2, matrix(c(-2, 0, 0)), c(1, 0, 25))
}
