% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_generate_sample.R
\name{norm_generate_sample}
\alias{norm_generate_sample}
\title{Generate data from a normal HMM}
\usage{
norm_generate_sample(num_sample, hmm, design, state_dep_dist_pooled = FALSE)
}
\arguments{
\item{num_sample}{The size of the desired sample (number of timesteps).}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{mu}, \code{sigma}, \code{beta},
\code{delta}.}

\item{design}{A list of design matrices for each subject with each row
indicating the time and each column indicating the value of the
covariate.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of the data and the states that generated the data.
}
\description{
This function generates data from a normal HMM with specified parameters.
}
\examples{
# define parameters
hmm <-  list(num_states = 3,
             num_variables = 2,
             num_subjects = 1,
             mu = list(matrix(c(12, 18, 22),
                       ncol = 3, nrow = 1, byrow = TRUE),
                       matrix(c(-12, -7, 0),
                              ncol = 3, nrow = 1, byrow = TRUE)),
             sigma = list(matrix(c(3, 1, 1.5),
                                 ncol = 3, nrow = 1, byrow = TRUE),
                          matrix(c(1, 3, 2),
                                 ncol = 3, nrow = 1, byrow = TRUE)),
             beta  = matrix(c(0.01, 0.02, 0.001,
                              0.01, 0.03, 0.004,
                              0.01, 0.01, 0.003,
                              0.01, 0.04, 0.002,
                              0.01, 0.01, 0.004,
                              0.01, 0.03, 0.001),
                              ncol = 3, nrow = 6, byrow = TRUE),
             delta = list(c(0.3, 0.2, 0.5)))

num_sample       <- 1000
design           <- list(matrix(0, nrow = 1000, ncol = 3))
design[[1]][, 1] <- 1 # First column is the intercept
design[[1]][, 2] <- sample(c(0, 1), size = 1000,
                           prob = c(0.3, 0.7), replace = TRUE)
design[[1]][, 3] <- rnorm(1000, mean = 5, sd = 1)

# generate sample
norm_generate_sample(num_sample, hmm, design, state_dep_dist_pooled = FALSE)
}
