% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ci.R
\name{norm_hist_ci}
\alias{norm_hist_ci}
\title{Title}
\usage{
norm_hist_ci(
  x,
  viterbi,
  conf_intervals,
  num_states,
  num_subjects,
  num_variables,
  hmm,
  state_dep_dist_pooled = FALSE,
  width = 1,
  n = 100,
  level = 0.975,
  x_step = 0.2
)
}
\arguments{
\item{x_step}{}
}
\value{

}
\description{
Title
}
