% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_loglikelihood.R
\name{norm_loglikelihood}
\alias{norm_loglikelihood}
\title{Compute negative log-likelihood of normal HMM parameters}
\usage{
norm_loglikelihood(
  working_params,
  x,
  design,
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{working_params}{A vector of the working normal parameters for the
HMM.}

\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{design}{A list of design matrices for each subject with each row
indicating the time and each column indicating the value of the
covariate.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A number indicating the negative loglikelihood
}
\description{
This function computes the negative log-likelihood that the given normal
HMM parameters could have generated the data being fit.
}
