% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_natural_params.R, R/norm_natural_params.R
\name{norm_natural_params}
\alias{norm_natural_params}
\title{Transform gamma HMM parameters from working to natural}
\usage{
norm_natural_params(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)

norm_natural_params(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability matrix depends on.}

\item{working_params}{A vector of the working normal parameters for the
HMM as outputted by \code{norm_working_params()}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of the natural paramters.

A list of the natural paramters.
}
\description{
This function transforms the working gamma HMM parameters back into the
original format of the natural parameters and outputs them as a list. This
function is the reverse of \code{norm_working_params()}.

This function transforms the working normal HMM parameters back into the
original format of the natural parameters and outputs them as a list. This
function is the reverse of \code{norm_working_params()}.
}
\examples{
### 1. Get the working parameters
# define values of parameters
num_states     <- 2
num_variables  <- 2
num_subjects   <- 2
num_covariates <- 2
mu             <- list(matrix(c(1, 5, 2, 4), 2, 2, byrow = TRUE),
                       matrix(c(1, 5, 2, 4), 2, 2, byrow = TRUE))
sigma          <- list(matrix(c(1, 2, 1, 1.5), 2, 2, byrow = TRUE),
                       matrix(c(1, 2, 1, 1.5), 2, 2, byrow = TRUE))
beta           <- matrix(c(-2, 0, 0), nrow = 1, ncol = 3)
delta          <- list(c(1/2, 1/2), c(1/2, 1/2))

#transform to working parametes
working_params <- norm_working_params(num_states, num_variables,
                                      num_subjects, mu, sigma, beta, delta)
### 2. Transform working_params to natural
norm_natural_params(num_states, num_variables, num_subjects, num_covariates,
                    working_params, state_dep_dist_pooled = FALSE)
}
