% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ci.R
\name{norm_natural_vec}
\alias{norm_natural_vec}
\title{Convert working parameters to a vector of natural parameters}
\usage{
norm_natural_vec(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  working_params,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects that generated the data.}

\item{num_covariates}{The number of covariates in the data that the
transition probability depends on.}

\item{working_params}{A vector of the working normal parameters for the
HMM.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A vector of the natural parameters.
}
\description{
Convert working parameters to a vector of natural parameters
}
