% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_viterbi.R
\name{norm_viterbi}
\alias{norm_viterbi}
\title{Global decoding by the Viterbi algorithm}
\usage{
norm_viterbi(x, hmm, state_dep_dist_pooled = FALSE)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{num_covariates}, \code{mu},
\code{sigma}, \code{gamma}, \code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A matrix with each column containing the sequence of states for the
given subject.
}
\description{
This function takes in data x assumed to be generated by the HMM \code{hmm} and
outputs the most likely sequence of states that could have generated the
data using global decoding by the Viterbi algorithm.
}
