% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_working_params.R
\name{norm_working_params}
\alias{norm_working_params}
\title{Transform normal HMM parameters from natural to working}
\usage{
norm_working_params(
  num_states,
  num_variables,
  num_subjects,
  mu,
  sigma,
  beta,
  delta
)
}
\arguments{
\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{mu}{A list of matrices containing the means of the state dependent
normal distribution. Each matrix corresponds to a different variable,
each row corresponds to a different subject and each column corresponds
to a different state.}

\item{sigma}{A list of matrices containing the standard deviations of the
state dependent normal distribution. Each matrix corresponds to a
different variable, each row corresponds to a different subject and each
column corresponds to a different state.}

\item{beta}{A matrix of regression coefficients for the effect of the
covariates on the transition probability matrix \code{gamma}.}

\item{delta}{A list of the initial state distributions for each subject.}
}
\value{
A single vector containing working parameters.
}
\description{
The function transforms the natural normal HMM parameters that have
additional constraints into working parameters that incorporate the
constraints.
}
\examples{
# define values of parameters
num_states    <- 2
num_variables <- 2
num_subjects  <- 2
mu            <- list(matrix(c(1, 5, 2, 4), 2, 2, byrow = TRUE),
                      matrix(c(1, 5, 2, 4), 2, 2, byrow = TRUE))
sigma         <- list(matrix(c(1, 2, 1, 1.5), 2, 2, byrow = TRUE),
                      matrix(c(1, 2, 1, 1.5), 2, 2, byrow = TRUE))
beta          <- matrix(c(-2, 0, 0), nrow = 1, ncol = 3)
delta         <- list(c(1/2, 1/2), c(1/2, 1/2))

#transform to working parametes
norm_working_params(num_states, num_variables, num_subjects,
                    mu, sigma, beta, delta)
}
