% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_analysis.R
\name{plot_tpm_entries}
\alias{plot_tpm_entries}
\title{Plot entries of transition probability matrices}
\usage{
plot_tpm_entries(
  num_states,
  covar_vec,
  cov_ci,
  covariate_name = "Temp - mean(Temp)"
)
}
\arguments{
\item{num_states}{The number of states in the HMM.}

\item{covar_vec}{The vector of covariate values to plot the entries over.}

\item{covariate_name}{Label for the x-axis.}

\item{tpm_entries_list}{A list of 3 matrices, one for the estimate and the
upper, and lower confidence interval, with the rows of the matrices
containing the entries of the transition probability matrices at each
covariate value.}
}
\value{
A list with a plot for each of the transition probability entries
varying with the covariate.
}
\description{
This function plots the transition probability matrix entries with their
confidence intervals for different covariate values.
}
