% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr_plotting.R
\name{psr_acf}
\alias{psr_acf}
\title{ACF plot of normal pseudo-residuals}
\usage{
psr_acf(forecast_psr, num_subjects)
}
\arguments{
\item{forecast_psr}{A list containing a vector of pseudo-residuals for each
subject.}

\item{num_subjects}{The number of subjects/trials that generated the data.}
}
\value{
An acf plot of the normal-pseudo residuals for each subject.
}
\description{
This function produces a ACF plot of the normal pseudo-residuals for each
subject.
}
