% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr_plotting.R
\name{psr_plot}
\alias{psr_plot}
\title{Plot normal pseudo-residuals}
\usage{
psr_plot(forcast_psr, num_subjects)
}
\arguments{
\item{forcast_psr}{A list containing a vector of pseudo-residuals for each
subject.}

\item{num_subjects}{The number of subjects/trials that generated the data.}
}
\value{
A plot of the normal-pseudo residuals for each subject.
}
\description{
This function produces a scatter plot of the normal pseudo-residuals in time
for each subject.
}
