% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr_plotting.R
\name{psr_qq}
\alias{psr_qq}
\title{Quantile-quantile plot of normal pseudo-residuals}
\usage{
psr_qq(forecast_psr, num_subjects)
}
\arguments{
\item{forecast_psr}{A list containing a vector of pseudo-residuals for each
subject.}

\item{num_subjects}{The number of subjects/trials that generated the data.}
}
\value{
A qq plot of the normal-pseudo residuals for each subject.
}
\description{
This function produces a quantile-quantile plot of the normal
pseudo-residuals for each subject.
}
