% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_vec}
\alias{split_vec}
\title{Split vector into list}
\usage{
split_vec(vector, start, end, length, exp = FALSE)
}
\arguments{
\item{vector}{The vector to be split into a list.}

\item{start}{The first index of the vector vector that will be included in
the list.}

\item{end}{The last index of the vector vector that will be included in the
list.}

\item{length}{A value indicating the length of the segments that vector
will be split into.}

\item{exp}{A logical variable indicating whether the elements of vector
should be transformed with \code{exp()}.}
}
\value{
A list of the sections of the split vector.
}
\description{
This function splits a vector from the index start to the index end into
segments of length length and complies them into a list. This function can
also apply the exponential function to the vector elements if desired.
}
\examples{
split_vec(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 6, 2)
split_vec(c(0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6), 2, 7, 3, exp = TRUE)
}
