% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_histogram_plotting.R
\name{timeseries_plot}
\alias{timeseries_plot}
\title{Plot timeseries of observations}
\usage{
timeseries_plot(
  x,
  states,
  num_subjects,
  num_variables,
  variable_names = c("Var 1", "Var 2", "Var 3"),
  subject_names = c("Subject 1", "Subject 2", "Subject 3", "Subject 4"),
  length = 300
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{states}{A matrix with the columns containing the sequence of states
that generated the data in \code{x} for the given subject.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_variables}{The number of variables in the data.}

\item{variable_names}{A vector containing the names of the variables in the
data \code{x}.}

\item{subject_names}{A vector containing the names of the subjects generating
the data \code{x}.}

\item{length}{A number indicating the desired length of the timeseries plot.}
}
\value{
A grid of plots of the time series for each subject and variable.
}
\description{
Plot timeseries of observations
}
