% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ci_functions.R
\name{norm_ci}
\alias{norm_ci}
\title{Compute the confidence intervals}
\usage{
norm_ci(
  hmm,
  state_dep_dist_pooled = FALSE,
  n = 100,
  level = 0.975,
  raw_sample = FALSE
)
}
\arguments{
\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects},\code{num_covariates}, \code{mu},
\code{sigma}, \code{beta}, \code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}

\item{n}{The number of samples in the Monte Carlo fitting.}

\item{level}{A number indicating the level of confidence for the desired
interval.}

\item{raw_sample}{A logical variable indicating wheter to simply output the
\code{n} sampled \code{mu}s and \code{sigma}s from the Monte Carlo estimate.}
}
\value{
Either a list of the sampled \code{mu}s and \code{sigma}s or the list of
confidence intervals for each parameter.
}
\description{
This function computes the confidence intervals for the estimated parameters
\code{mu}, \code{sigma}, and \code{beta} using the variances outputted from \code{norm_fit_hmm()}
and the Monte Carlo method of estimation.
}
