% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confindence_interval_functions.R
\name{norm_ci_data}
\alias{norm_ci_data}
\title{Reformat confidence interval data}
\usage{
norm_ci_data(
  num_states,
  num_variables,
  num_subjects,
  num_covariates,
  estimate_vec,
  upper_vec,
  lower_vec,
  state_dep_dist_pooled = FALSE
)
}
\arguments{
\item{estimate_vec}{A vector containing the estimated natural parameters.}

\item{upper_vec}{A vector containing the upper confidence interval of the
estimated natural parameters.}

\item{lower_vec}{A vector containing the lower confidence interval of the
estimated natural parameters.}
}
\value{
A list containing the upper and lower confidence interval for each
estimated parameter.
}
\description{
Reformat confidence interval data
}
