% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confindence_interval_functions.R
\name{norm_dist_ci_data}
\alias{norm_dist_ci_data}
\title{#' Compute the confidence intervals of fitted normal distributions}
\usage{
norm_dist_ci_data(
  x,
  num_states,
  num_variables,
  num_subjects,
  sample,
  state_dep_dist_pooled = FALSE,
  x_step = 0.2,
  n = 100,
  level = 0.975
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects that generated the data.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}

\item{x_step}{A value indicating the step length for the range of
observation values.}

\item{n}{The number of samples in the Monte Carlo fitting.}

\item{level}{A number indicating the level of confidence for the desired
interval.}

\item{sample_params}{A matrix of the natural parameters.}
}
\value{

}
\description{
This function computes the confidence intervals for the fitted normal state
dependent distributions using the Monte Carlo approach.
}
