% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_generate_sample.R
\name{norm_generate_sample}
\alias{norm_generate_sample}
\title{Generate data from a normal HMM}
\usage{
norm_generate_sample(num_sample, hmm, design, state_dep_dist_pooled = FALSE)
}
\arguments{
\item{num_sample}{The size of the desired sample (number of timesteps).}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{num_covariates}, \code{mu},
\code{sigma}, \code{beta}, \code{delta}.}

\item{design}{A list of design matrices for each subject with each row
indicating the time and each column indicating the value of the
covariate.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}
}
\value{
A list of the data and the states that generated the data.
}
\description{
This function generates data from a normal HMM with specified parameters.
}
