% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_histogram_plotting.R
\name{norm_hist}
\alias{norm_hist}
\title{Plot histograms}
\usage{
norm_hist(
  sample,
  num_states,
  num_variables,
  num_subjects,
  hmm,
  width = 1,
  x_step = 0.2
)
}
\arguments{
\item{sample}{The sample generated by \code{norm_generate_sample()}.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_variables}{The number of variables in the data.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{mu}, \code{sigma}, \code{gamma},
\code{delta}.}

\item{width}{The width of the histogram bins.}

\item{x_step}{A value indicating the step length for the range of
observation values.}
}
\value{
Histograms of the data with the distributions overlayed.
}
\description{
This function plots the histograms for each subject and variable with the
true state dependent distributions overlayed.
}
