% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_histogram_plotting.R
\name{norm_hist}
\alias{norm_hist}
\title{Title}
\usage{
norm_hist(
  sample,
  num_states,
  num_subjects,
  num_variables,
  hmm,
  width = 1,
  x_step = 0.2
)
}
\arguments{
\item{x_step}{}
}
\value{

}
\description{
Title
}
