% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval_functions.R
\name{norm_hist_ci}
\alias{norm_hist_ci}
\title{Title}
\usage{
norm_hist_ci(
  x,
  viterbi,
  num_states,
  num_subjects,
  num_variables,
  hmm,
  state_dep_dist_pooled = FALSE,
  width = 1,
  n = 100,
  level = 0.975,
  x_step = 0.2
)
}
\arguments{
\item{x}{The data to be fit with an HMM in the form of a 3D array. The
first index (row) corresponds to time, the second (column) to the
variable number, and the third (matrix number) to the subject number.}

\item{num_states}{The number of states in the desired HMM.}

\item{num_subjects}{The number of subjects/trials that generated the data.}

\item{num_variables}{The number of variables in the data.}

\item{hmm}{A list of parameters that specify the normal HMM, including
\code{num_states}, \code{num_variables}, \code{num_subjects}, \code{mu}, \code{sigma}, \code{gamma},
\code{delta}.}

\item{state_dep_dist_pooled}{A logical variable indiacting whether the
state dependent distribution parameters \code{mu} and \code{sigma} should be
treated as equal for all subjects.}

\item{width}{The width of the histogram bins.}

\item{n}{The number of samples in the Monte Carlo fitting.}

\item{level}{A number indicating the level of confidence for the desired
interval.}

\item{x_step}{A value indicating the step length for the range of
observation values.}
}
\value{

}
\description{
Description
}
